Add buildSrc convention plugin and use configuration caching
