+++ bukkit/build.gradle.kts
@@ -1,3 +1,7 @@
+plugins {
+    id("common")
+}
+
 repositories {
     maven("https://hub.spigotmc.org/nexus/content/repositories/snapshots/")
     maven("https://repo.papermc.io/repository/maven-public/")
@@ -5,11 +9,11 @@ repositories {
 }
 
 dependencies {
-    compileOnly(group = "org.spigotmc", name = "spigot-api", version = "1.21.2-R0.1-SNAPSHOT")
-    compileOnly(group = "com.github.Puremin0rez", name = "WorldBorder", version = "1.19") {
+    compileOnly("org.spigotmc:spigot-api:1.21.2-R0.1-SNAPSHOT")
+    compileOnly("com.github.Puremin0rez:WorldBorder:1.19") {
         isTransitive = false
     }
-    implementation(group = "org.bstats", name = "bstats-bukkit", version = "3.0.2")
+    implementation("org.bstats:bstats-bukkit:3.0.2")
     implementation(project(":chunky-common"))
     implementation(project(":chunky-paper"))
     implementation(project(":chunky-folia"))
@@ -17,14 +21,15 @@ dependencies {
 
 tasks {
     processResources {
+        val props = mapOf(
+            "name" to project.property("artifactName")!!,
+            "version" to project.version,
+            "group" to project.group,
+            "author" to project.property("author")!!,
+            "description" to project.property("description")!!,
+        )
         filesMatching("plugin.yml") {
-            expand(
-                "name" to project.property("artifactName")!!,
-                "version" to project.version,
-                "group" to project.group,
-                "author" to project.property("author")!!,
-                "description" to project.property("description")!!,
-            )
+            expand(props)
         }
     }
     shadowJar {
