+++ common/build.gradle.kts
@@ -1,16 +1,19 @@
+plugins {
+    id("common")
+}
+
 dependencies {
-    compileOnly(group = "com.google.code.gson", name = "gson", version = "2.8.9")
-    testImplementation(group = "junit", name = "junit", version = "4.13.2")
-    testImplementation(group = "com.google.code.gson", name = "gson", version = "2.10.1")
+    compileOnly("com.google.code.gson:gson:2.8.9")
+    testImplementation("junit:junit:4.13.2")
+    testImplementation("com.google.code.gson:gson:2.10.1")
     implementation(project(":chunky-nbt"))
 }
 
 tasks {
     processResources {
+        val props = mapOf("version" to project.version)
         filesMatching("version.properties") {
-            expand(
-                "version" to project.version
-            )
+            expand(props)
         }
     }
     javadoc {
@@ -19,7 +22,7 @@ tasks {
                 allJava
             }
         }
-        setDestinationDir(rootProject.projectDir.resolve("docs/chunky/javadoc"))
+        destinationDir = rootProject.projectDir.resolve("docs/chunky/javadoc")
         include("org/popcraft/chunky/api/**")
         exclude("org/popcraft/chunky/api/ChunkyAPIImpl.java")
     }
