+++ fabric/build.gradle.kts
@@ -1,30 +1,32 @@
 plugins {
+    id("common")
     id("dev.architectury.loom") version "1.13-SNAPSHOT"
 }
 
 val shade: Configuration by configurations.creating
 
 dependencies {
-    minecraft(group = "com.mojang", name = "minecraft", version = "1.21.11")
+    minecraft("com.mojang:minecraft:1.21.11")
     mappings(loom.officialMojangMappings())
-    modImplementation(group = "net.fabricmc", name = "fabric-loader", version = "0.18.1")
-    modImplementation(group = "net.fabricmc.fabric-api", name = "fabric-api", version = "0.139.4+1.21.11")
-    modCompileOnly(group = "me.lucko", name = "fabric-permissions-api", version = "0.6.1")
+    modImplementation("net.fabricmc:fabric-loader:0.18.1")
+    modImplementation("net.fabricmc.fabric-api:fabric-api:0.139.4+1.21.11")
+    modCompileOnly("me.lucko:fabric-permissions-api:0.6.1")
     implementation(project(":chunky-common"))
     shade(project(":chunky-common"))
 }
 
 tasks {
     processResources {
+        val props = mapOf(
+            "id" to rootProject.name,
+            "version" to project.version,
+            "name" to project.property("artifactName")!!,
+            "description" to project.property("description")!!,
+            "author" to project.property("author")!!,
+            "github" to project.property("github")!!
+        )
         filesMatching("fabric.mod.json") {
-            expand(
-                "id" to rootProject.name,
-                "version" to project.version,
-                "name" to project.property("artifactName")!!,
-                "description" to project.property("description")!!,
-                "author" to project.property("author")!!,
-                "github" to project.property("github")!!
-            )
+            expand(props)
         }
     }
     shadowJar {
