+++ folia/build.gradle.kts
@@ -1,7 +1,11 @@
+plugins {
+    id("common")
+}
+
 repositories {
     maven("https://repo.papermc.io/repository/maven-public/")
 }
 
 dependencies {
-    compileOnly(group = "dev.folia", name = "folia-api", version = "1.20.2-R0.1-SNAPSHOT")
+    compileOnly("dev.folia:folia-api:1.20.2-R0.1-SNAPSHOT")
 }
