+++ forge/build.gradle.kts
@@ -1,13 +1,14 @@
 plugins {
+    id("common")
     id("dev.architectury.loom") version "1.13-SNAPSHOT"
 }
 
 val shade: Configuration by configurations.creating
 
 dependencies {
-    minecraft(group = "com.mojang", name = "minecraft", version = "1.21.11")
+    minecraft("com.mojang:minecraft:1.21.11")
     mappings(loom.officialMojangMappings())
-    forge(group = "net.minecraftforge", name = "forge", version = "1.21.11-61.0.2")
+    forge("net.minecraftforge:forge:1.21.11-61.0.2")
     implementation(project(":chunky-common"))
     shade(project(":chunky-common"))
 }
@@ -20,15 +21,16 @@ loom {
 
 tasks {
     processResources {
+        val props = mapOf(
+            "github" to project.property("github")!!,
+            "id" to rootProject.name,
+            "version" to project.version,
+            "name" to project.property("artifactName")!!,
+            "author" to project.property("author")!!,
+            "description" to project.property("description")!!
+        )
         filesMatching("META-INF/mods.toml") {
-            expand(
-                "github" to project.property("github")!!,
-                "id" to rootProject.name,
-                "version" to project.version,
-                "name" to project.property("artifactName")!!,
-                "author" to project.property("author")!!,
-                "description" to project.property("description")!!
-            )
+            expand(props)
         }
     }
     jar {
