+++ gradle.properties
@@ -1,7 +1,11 @@
-org.gradle.jvmargs=-Xmx2G
 group=org.popcraft
 artifactName=Chunky
 version=1.4
 description=Pre-generates chunks, quickly, efficiently, and safely
 github=https://github.com/pop4959/Chunky
 author=pop4959
+
+org.gradle.configuration-cache=true
+org.gradle.parallel=true
+org.gradle.caching=true
+org.gradle.jvmargs=-Xmx2G
